git-svn-id: https://r7/svn/WebGL@170 45b56d4a-f8b8-b242-b192-c14f84797ed6

+++ Shaders/FluidSim/3D/advection_pass.fs.glsl
@@ -2,8 +2,8 @@
 // #extension GL_EXT_shader_texture_lod : require
 //Advekcija brzine, (jedan pass), nakon ovog moze ici MacCormack korekcija. input je brzina nakon viscosity, output je advected brzina
 
-precision mediump float;
-precision mediump sampler3D;
+precision highp float;
+precision highp sampler3D;
 
 #global_defines
 #include "defines"
